[backend] Store from & to representative in relationship to boost performance
