+++ opencti-platform/opencti-graphql/src/database/middleware.js
@@ -21,8 +21,6 @@ import {
   computeAverage,
   extractIdsFromStoreObject,
   fillTimeSeries,
-  generateCreateMessage,
-  generateUpdateMessage,
   INDEX_INFERRED_RELATIONSHIPS,
   inferIndexFromConceptType,
   isEmptyField,
@@ -236,6 +234,7 @@ import { schemaRelationsRefDefinition } from '../schema/schema-relationsRef';
 import { validateInputCreation, validateInputUpdate } from '../schema/schema-validator';
 import { getMandatoryAttributesForSetting } from '../domain/attribute';
 import { telemetry } from '../config/tracing';
+import { generateCreateMessage, generateUpdateMessage } from './generate-message';
 
 // region global variables
 export const MAX_BATCH_SIZE = 300;
