+++ opencti-platform/opencti-graphql/src/database/utils.js
@@ -1,15 +1,15 @@
 import * as R from 'ramda';
 import moment from 'moment';
-import { DatabaseError, UnsupportedError } from '../config/errors';
+import { DatabaseError } from '../config/errors';
 import { isHistoryObject, isInternalObject } from '../schema/internalObject';
 import { isStixMetaObject } from '../schema/stixMetaObject';
 import { isStixDomainObject } from '../schema/stixDomainObject';
-import { ENTITY_HASHED_OBSERVABLE_STIX_FILE, isStixCyberObservable } from '../schema/stixCyberObservable';
+import { isStixCyberObservable } from '../schema/stixCyberObservable';
 import { isInternalRelationship } from '../schema/internalRelationship';
 import { isStixCoreRelationship } from '../schema/stixCoreRelationship';
 import { isStixSightingRelationship } from '../schema/stixSightingRelationship';
-import { isStixObject } from '../schema/stixCoreObject';
 import conf from '../config/conf';
+import { now } from '../utils/format';
 import { FROM_START_STR, now, UNTIL_END_STR } from '../utils/format';
 import { isBasicRelationship } from '../schema/stixRelationship';
 import { truncate } from '../utils/mailData';
@@ -17,7 +17,6 @@ import { isDateAttribute, isDictionaryAttribute, isJsonAttribute,isObjectAttribu
 import { schemaRelationsRefDefinition } from '../schema/schema-relationsRef';
 import { creators } from '../schema/attribute-definition';
 import { isStixRefRelationship } from '../schema/stixRefRelationship';
-import { extractEntityRepresentativeName } from './entity-representative';
 
 export const ES_INDEX_PREFIX = conf.get('elasticsearch:index_prefix') || 'opencti';
 const rabbitmqPrefix = conf.get('rabbitmq:queue_prefix');
@@ -246,89 +245,6 @@ export const inferIndexFromConceptType = (conceptType, inferred = false) => {
   if (isStixRefRelationship(conceptType)) return INDEX_STIX_META_RELATIONSHIPS;
 
   throw DatabaseError(`Cant find index for type ${conceptType}`);
-};
-
-export const generateMergeMessage = (instance, sources) => {
-  const name = extractEntityRepresentativeName(instance);
-  const sourcesNames = sources.map((source) => extractEntityRepresentativeName(source)).join(', ');
-  return `merges ${instance.entity_type} \`${sourcesNames}\` in \`${name}\``;
-};
-const generateCreateDeleteMessage = (type, instance) => {
-  const name = extractEntityRepresentativeName(instance);
-  if (isStixObject(instance.entity_type)) {
-    let entityType = instance.entity_type;
-    if (entityType === ENTITY_HASHED_OBSERVABLE_STIX_FILE) {
-      entityType = 'File';
-    }
-    return `${type}s a ${entityType} \`${name}\``;
-  }
-  if (isBasicRelationship(instance.entity_type)) {
-    const from = extractEntityRepresentativeName(instance.from);
-    let fromType = instance.from.entity_type;
-    if (fromType === ENTITY_HASHED_OBSERVABLE_STIX_FILE) {
-      fromType = 'File';
-    }
-    const to = extractEntityRepresentativeName(instance.to);
-    let toType = instance.to.entity_type;
-    if (toType === ENTITY_HASHED_OBSERVABLE_STIX_FILE) {
-      toType = 'File';
-    }
-    return `${type}s the relation ${instance.entity_type} from \`${from}\` (${fromType}) to \`${to}\` (${toType})`;
-  }
-  return '-';
-};
-
-export const generateCreateMessage = (instance) => {
-  return generateCreateDeleteMessage(EVENT_TYPE_CREATE, instance);
-};
-export const generateDeleteMessage = (instance) => {
-  return generateCreateDeleteMessage(EVENT_TYPE_DELETE, instance);
-};
-
-export const generateUpdateMessage = (entityType, inputs) => {
-  const inputsByOperations = R.groupBy((m) => m.operation ?? UPDATE_OPERATION_REPLACE, inputs);
-  const patchElements = Object.entries(inputsByOperations);
-  if (patchElements.length === 0) {
-    throw UnsupportedError('[OPENCTI] Error generating update message with empty inputs');
-  }
-  // noinspection UnnecessaryLocalVariableJS
-  const generatedMessage = patchElements.slice(0, 3).map(([type, operations]) => {
-    return `${type}s ${operations.slice(0, 3).map(({ key, value }) => {
-      let message = 'nothing';
-      let convertedKey;
-      const relationsRefDefinition = schemaRelationsRefDefinition.getRelationRef(entityType, key);
-      if (relationsRefDefinition) {
-        convertedKey = relationsRefDefinition.label ?? relationsRefDefinition.stixName;
-      } else {
-        const attributeDefinition = schemaAttributesDefinition.getAttribute(entityType, key);
-        convertedKey = attributeDefinition.label ?? attributeDefinition.name;
-      }
-      const fromArray = Array.isArray(value) ? value : [value];
-      const values = fromArray.slice(0, 3).filter((v) => isNotEmptyField(v));
-      if (isNotEmptyField(values)) {
-        // If update is based on internal ref, we need to extract the value
-        if (relationsRefDefinition) {
-          message = values.map((val) => truncate(extractEntityRepresentativeName(val))).join(', ');
-        } else if (key === creators.name) {
-          message = 'itself'; // Creator special case
-        } else if (isDictionaryAttribute(key)) {
-          message = Object.entries(R.head(values)).map(([k, v]) => truncate(`${k}:${v}`)).join(', ');
-        } else if (isObjectAttribute(key)) {
-          message = 'elements';
-        } else if (isJsonAttribute(key)) {
-          message = values.map((v) => truncate(JSON.stringify(v)));
-        } else if (isDateAttribute(key)) {
-          message = values.map((v) => ((v === FROM_START_STR || v === UNTIL_END_STR) ? 'nothing' : v));
-        } else {
-          // If standard primitive data, just join the values
-          message = values.join(', ');
-        }
-      }
-      return `\`${message}\` in \`${convertedKey}\`${(fromArray.length > 3) ? ` and ${fromArray.length - 3} more items` : ''}`;
-    }).join(' - ')}`;
-  }).join(' | ');
-  // Return generated update message
-  return `${generatedMessage}${patchElements.length > 3 ? ` and ${patchElements.length - 3} more operations` : ''}`;
 };
 
 export const pascalize = (s) => {
