+++ opencti-platform/opencti-graphql/src/rules/containerWithRefsBuilder.ts
@@ -14,7 +14,6 @@ import { STIX_EXT_OCTI } from '../types/stix-extensions';
 import { internalFindByIds, internalLoadById, listAllRelations } from '../database/middleware-loader';
 import type { RelationCreation, UpdateEvent } from '../types/event';
 import {
-  generateUpdateMessage,
   READ_DATA_INDICES,
   UPDATE_OPERATION_ADD,
   UPDATE_OPERATION_REMOVE
@@ -23,6 +22,7 @@ import type { AuthContext } from '../types/user';
 import { executionContext, RULE_MANAGER_USER } from '../utils/access';
 import { buildStixUpdateEvent, publishStixToStream } from '../database/redis';
 import { INPUT_DOMAIN_TO, INPUT_OBJECTS, RULE_PREFIX } from '../schema/general';
+import { generateUpdateMessage } from '../database/generate-message';
 
 const buildContainerRefsRule = (ruleDefinition: RuleDefinition, containerType: string, relationTypes: RelationTypes): RuleRuntime => {
   const { id } = ruleDefinition;
